requirements.txt
streamlit
opencv-python-headless
numpy
Pillow
pyyaml
torch==1.13.1
torchvision==0.14.1
detectron2 @ git+https://github.com/facebookresearch/detectron2.git